+++ src/indices/spent_index.h
